Change VnodeDiskObject.inodeNumber to a union of a pointer to the
RVM representation for directories and an container file identifying
integer for files and symlinks. Initializing realm::realmid with a
global counter value, the 32-bit truncated pointer is not a useful
uniquifier on a 64-bit system. And various other cleanups.

+++ coda-src/util/bstree.cc
@@ -428,7 +428,7 @@ bstree_iterator::bstree_iterator(bstree& Bst, BstIterOrder Order) {
 bsnode *bstree_iterator::operator()() {
     bsnode *prev = cbsnode;
 
-    switch((unsigned int)cbsnode) {
+    switch((intptr_t)cbsnode) {
 	case -1:		/* state == NOTSTARTED */
 	    if (order == BstAscending) {
 		cbsnode = cbstree->first();
